use int16_t

+++ src/WBReceiver.cpp
@@ -232,7 +232,7 @@ void WBReceiver::processPacket(const uint8_t wlan_idx, const pcap_pkthdr &hdr, c
         //m_console->debug("x_n_missing_packets:{} x_n_received_packets:{} n_total_packets:{}",x_n_missing_packets,x_n_received_packets,n_total_packets);
         if(n_total_packets>=1){
           const double loss_perc=static_cast<double>(x_n_missing_packets)/static_cast<double>(n_total_packets)*100.0;
-          x_curr_packet_loss_perc=static_cast<int>(std::lround(loss_perc));
+          x_curr_packet_loss_perc=static_cast<int16_t>(std::lround(loss_perc));
           //m_console->debug("Packet loss:{} % {} %",x_curr_packet_loss_perc,loss_perc);
         }
         x_curr_n_of_big_gaps=x_n_big_gaps_since_last;
